change names of constants
